Move stats-related selectors to state/stats.
